refac : 관리자용 셀프 소개 레포지토리 분리.
